% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{aoi_size}
\alias{aoi_size}
\title{Compute AOI Size}
\usage{
aoi_size(jsonfile)
}
\arguments{
\item{jsonfile}{path to geojson file}
}
\value{
'units' object with area in km^2
}
\description{
Check the size of an AOi in km^2
}
