% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_orders}
\alias{check_orders}
\title{Check All Orders}
\usage{
check_orders()
}
\value{
a data frame of orders containing the time created, order ID, state, and name
}
\description{
Get a table of the most recent orders made
}
