% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_order}
\alias{download_order}
\title{Download Zip Bundle}
\usage{
download_order(order_id, ddir = "data")
}
\arguments{
\item{order_id}{the Planet order ID}

\item{ddir}{directory into which file should be downloaded}
}
\description{
Download files associated with an order ID
}
