% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_all_ids}
\alias{get_all_ids}
\title{Get All IDs from Search}
\usage{
get_all_ids(r)
}
\arguments{
\item{Result}{from search query}
}
\value{
a character vector of IDs
}
\description{
Follow _next URLs to get all IDs when spread across pages
}
