% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plnt_order}
\alias{plnt_order}
\title{Order Planet API}
\usage{
plnt_order(ord)
}
\arguments{
\item{ord}{R object containing the details of the order, such as filters, delivery, or notifications}
}
\value{
A list containing the order results
}
\description{
Post an order query to the Planet Labs API
}
