% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plnt_plot_any}
\alias{plnt_plot_any}
\title{Plot Image}
\usage{
plnt_plot_any(x, pch = 15, ...)
}
\arguments{
\item{x}{a GDAL object}

\item{pch}{pch for plot (default 15)}

\item{...}{other parameters passed to \code{plot}}
}
\description{
Normalize color bands in RGB and then combine values in red (band3), green (band2), and blue (band1) channels to create a "color" image
}
\details{
Plot Any Satellite Image via RGB
}
