% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plnt_plot_visual}
\alias{plnt_plot_visual}
\title{Plot Visual Image}
\usage{
plnt_plot_visual(x, ...)
}
\arguments{
\item{x}{a GDAL object}

\item{...}{other parameters passed to \code{plot}}

\item{pch}{pch for plot (default 15)}
}
\description{
Combine values in red (band3), green (band2), and blue (band1) channels to create a "color" image
}
\details{
Plot a Visual Satellite Image
}
