% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plnt_quota}
\alias{plnt_quota}
\title{Check Quota}
\usage{
plnt_quota()
}
\value{
A data frame with columns 'quota_used', 'quota_sqkm', and 'pct_used'
}
\description{
Check your current download quota and how much of it you have used. The quota is measured in square kilometers
}
\details{
Check Current Download Quota
}
