% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plnt_search}
\alias{plnt_search}
\title{Search Planet API}
\usage{
plnt_search(searchq)
}
\arguments{
\item{searchq}{Search query represented as an R object}
}
\value{
An object containing the search results
}
\description{
Post a search query to the Planet Labs API
}
