% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plnt_status}
\alias{plnt_status}
\title{Check Status of an Order}
\usage{
plnt_status(order_id, silent = FALSE)
}
\arguments{
\item{order_id}{Order ID}

\item{silent}{should status summary be printed?}
}
\description{
Return the status information for an order ID
}
