% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plnt_plot_any}
\alias{plnt_plot_any}
\title{Plot Image}
\usage{
plnt_plot_any(x, mult = 1, shift = 0)
}
\arguments{
\item{x}{a GDAL object}

\item{mult}{scale factor for normalizing color bands}
}
\description{
Normalize color bands in RGB and then combine values in red (band3), green (band2), and blue (band1) channels to create a "color" image
}
\details{
Plot Any Satellite Image via RGB
}
