% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plnt_search}
\alias{plnt_search}
\title{Search Planet API}
\usage{
plnt_search(details_json)
}
\arguments{
\item{details_json}{path to JSON file containing the details of the search, such as filters}
}
\value{
A list containing the search results
}
\description{
Post a search query to the Planet Labs API
}
